shift age warning logic onto the server
