--- frontend/lib/articleAgeWarning.ts
@@ -1,41 +1,0 @@
-export const getAgeWarning = (webPublicationDate: Date): string => {
-    const warnLimitDays = 30;
-    const currentDate = new Date();
-    const dateThreshold = new Date();
-
-    dateThreshold.setDate(currentDate.getDate() - warnLimitDays);
-
-    const publicationDate = new Date(webPublicationDate);
-
-    // if the publication date is before the date threshold generate message
-    if (publicationDate < dateThreshold) {
-        // Unary + coerces dates to numbers for TypeScript
-        const diffMilliseconds = +currentDate - +publicationDate;
-        const diffSeconds = Math.floor(diffMilliseconds / 1000);
-        const diffMinutes = diffSeconds / 60;
-        const diffHours = diffMinutes / 60;
-        const diffDays = diffHours / 24;
-        const diffMonths = diffDays / 31;
-        const diffYears = diffDays / 365;
-
-        const message = 'This article is over';
-
-        if (diffYears >= 2) {
-            return `${message} ${Math.floor(diffYears)} years old`;
-        }
-
-        if (diffYears > 1) {
-            return `${message} 1 year old`;
-        }
-
-        if (diffMonths >= 2) {
-            return `${message} ${Math.floor(diffMonths)} months old`;
-        }
-
-        if (diffMonths > 1) {
-            return `${message} 1 month old`;
-        }
-    }
-
-    return '';
-};
