+++ frontend/lib/parse-capi/index.ts
@@ -92,32 +92,87 @@ const getLink = (data: {}, { isPillar }: { isPillar: boolean }): LinkType => ({
     mobileOnly: false,
 });
 
+const getAgeWarning = (webPublicationDate: Date): string | void => {
+    const warnLimitDays = 30;
+    const currentDate = new Date();
+    const dateThreshold = new Date();
+
+    dateThreshold.setDate(currentDate.getDate() - warnLimitDays);
+
+    const publicationDate = new Date(webPublicationDate);
+
+    // if the publication date is before the date threshold generate message
+    if (publicationDate < dateThreshold) {
+        // Unary + coerces dates to numbers for TypeScript
+        const diffMilliseconds = +currentDate - +publicationDate;
+        const diffSeconds = Math.floor(diffMilliseconds / 1000);
+        const diffMinutes = diffSeconds / 60;
+        const diffHours = diffMinutes / 60;
+        const diffDays = diffHours / 24;
+        const diffMonths = diffDays / 31;
+        const diffYears = diffDays / 365;
+
+        const message = 'This article is over';
+
+        if (diffYears >= 2) {
+            return `${message} ${Math.floor(diffYears)} years old`;
+        }
+
+        if (diffYears > 1) {
+            return `${message} 1 year old`;
+        }
+
+        if (diffMonths >= 2) {
+            return `${message} ${Math.floor(diffMonths)} months old`;
+        }
+
+        if (diffMonths > 1) {
+            return `${message} 1 month old`;
+        }
+    }
+};
+
 // TODO really it would be nice if we passed just the data we needed and
 // didn't have to do the transforms/lookups below. (While preserving the
 // validation on types.)
-export const extractArticleMeta = (data: {}): CAPIType => ({
-    headline: apply(
-        getNonEmptyString(data, 'config.page.headline'),
-        clean,
-        curly,
-    ),
-    standfirst: apply(
-        getNonEmptyString(data, 'contentFields.fields.standfirst'),
-        clean,
-        bigBullets,
-    ),
-    main: apply(getNonEmptyString(data, 'contentFields.fields.main'), clean),
-    body: getArray(data, 'contentFields.fields.blocks.body')
-        .map(block => block.bodyHtml)
-        .filter(Boolean)
-        .join(''),
-    author: getNonEmptyString(data, 'config.page.author'),
-    webPublicationDate: new Date(
+export const extractArticleMeta = (data: {}): CAPIType => {
+    const webPublicationDate = new Date(
         getNumber(data, 'config.page.webPublicationDate'),
-    ),
-    sectionName: getNonEmptyString(data, 'config.page.section'),
-    pageId: getNonEmptyString(data, 'config.page.pageId'),
-});
+    );
+
+    const articleMeta: CAPIType = {
+        webPublicationDate,
+        headline: apply(
+            getNonEmptyString(data, 'config.page.headline'),
+            clean,
+            curly,
+        ),
+        standfirst: apply(
+            getNonEmptyString(data, 'contentFields.fields.standfirst'),
+            clean,
+            bigBullets,
+        ),
+        main: apply(
+            getNonEmptyString(data, 'contentFields.fields.main'),
+            clean,
+        ),
+        body: getArray(data, 'contentFields.fields.blocks.body')
+            .map(block => block.bodyHtml)
+            .filter(Boolean)
+            .join(''),
+        author: getNonEmptyString(data, 'config.page.author'),
+        sectionName: getNonEmptyString(data, 'config.page.section'),
+        pageId: getNonEmptyString(data, 'config.page.pageId'),
+    };
+
+    const ageWarning = getAgeWarning(webPublicationDate);
+
+    if (ageWarning) {
+        articleMeta.ageWarning = ageWarning;
+    }
+
+    return articleMeta;
+};
 
 export const extractNavMeta = (data: {}): NavType => {
     let pillars = getArray(data, 'config.nav.pillars');
